Fix bug that caused w/h miscalculation.
I was calculating min X/Y with a maximum value of 0. That is, minX could never
be >0. Somehow this made it possible for the shrink check to work by only
checking width, when in fact it should've needed to check width and height.
Fixed that, started checking height too, and got rid of the unnecessary map
building, doing it only when we're printing. That made the program much much
faster.